bigInteger for phone number

+++ database/migrations/2021_06_08_162632_create_app_users_table.php
@@ -15,7 +15,7 @@ class CreateAppUsersTable extends Migration
     {
         Schema::create('app_users', function (Blueprint $table) {
             $table->id();
-            $table->integer('phone_number')->unique();
+            $table->bigInteger('phone_number')->unique();
             $table->string('email')->unique();
             $table->string('password');
             $table->rememberToken();
